refactor: mv walletconnect to provider/meta
