+++ src/provider/index.ts
@@ -1,1 +1,2 @@
+export * from './meta'
 export * from './signing'
