+++ src/provider/signing.ts
@@ -2,7 +2,7 @@ import type { TypedDataDomain, TypedDataField } from '@ethersproject/abstract-si
 import { _TypedDataEncoder } from '@ethersproject/hash'
 import type { JsonRpcProvider, JsonRpcSigner } from '@ethersproject/providers'
 
-import { getPeerMeta } from '../walletconnect'
+import { getPeerMeta } from './meta'
 
 // These are wallets which do not implement eth_signTypedData_v4, but *do* implement eth_signTypedData.
 // They are special-cased so that signing will still use EIP-712 (which is safer for the user).
